requirements.txt


# Core Azure Services
azure-search-documents==11.4.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# LangChain and AI
langchain==0.1.0
langchain-community==0.0.13
langchain-anthropic==0.1.1
anthropic==0.18.1

# Embeddings
sentence-transformers==2.3.1
transformers==4.37.2
torch==2.1.2

# PDF Processing
pypdf==4.0.1
pypdf2==3.0.1

# Text Processing
tiktoken==0.5.2

# Environment Management
python-dotenv==1.0.1

# Azure Functions (if deploying to Azure Functions)
azure-functions==1.18.0

# API/Web Framework (if building REST API)
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3

# Utilities
requests==2.31.0
numpy==1.26.3
pandas==2.2.0

# Development Tools
pytest==8.0.0
black==24.1.1
flake8==7.0.0